move files, update structure
