+++ src/index.js
@@ -1,13 +1,15 @@
-import { refs } from './js/reference.js';
-import { onTopArrow } from './js/lift-up.js';
-import { onSwitch } from './js/switch.js';
-import { renderFilms } from './js/renderFilms.js';
-import { apiService } from './js/api-service.js';
-import { toggleBackdrop } from './js/backdrop.js';
-import { renderPagination } from './js/pagination.js';
-import { spinnerOn } from './js/spinner.js';
-import { spinnerOff } from './js/spinner.js';
-import { onSmoothScroll } from './js/smoothScroll.js';
+import { refs } from './js/references/reference.js';
+import { onTopArrow } from './js/components/lift-up.js';
+import { fillModal } from './js/components/modal/fillModal.js';
+import { onSwitch } from './js/components/switch.js';
+import { onPaginationClick } from './js/pagination/onPaginationClick.js';
+import { renderFilms } from './js/render/renderFilms.js';
+import { apiService } from './js/API/api-service.js';
+import { toggleBackdrop } from './js/components/modal/backdrop.js';
+import { renderPagination } from './js/pagination/pagination.js';
+import { spinnerOn } from './js/components/spinner.js';
+import { spinnerOff } from './js/components/spinner.js';
+import { onSmoothScroll } from './js/components/smoothScroll.js';
 
 onTopArrow();
 onSwitch();
@@ -19,6 +21,7 @@ refs.form.addEventListener('submit', onFormSubmit);
 refs.paginationControls.addEventListener('click', onPaginationClick);
 refs.contentList.addEventListener('click', onGetInfoClick);
 
+//finction fetch popular films
 apiService.fetchFilms().then(({ data }) => {
   if (!apiService.allGenres) {
     apiService.fetchGenres();
@@ -28,44 +31,10 @@ apiService.fetchFilms().then(({ data }) => {
 
   renderPagination(apiService.page, apiService.totalPages);
 });
-
-function onPaginationClick(e) {
-  if (e.target.nodeName !== 'BUTTON') {
-    return;
-  }
-  if (e.target.classList.contains('js-next')) {
-    renderContentByPagination((apiService.page += 1));
-    return;
-  }
-  if (e.target.classList.contains('js-previous')) {
-    renderContentByPagination((apiService.page -= 1));
-    return;
-  }
-  if (e.target.classList.contains('js-second-previous')) {
-    renderContentByPagination((apiService.page -= 2));
-    return;
-  }
-  if (e.target.classList.contains('js-second-next')) {
-    renderContentByPagination((apiService.page += 2));
-    return;
-  }
-  if (e.target.classList.contains('js-first')) {
-    renderContentByPagination(1);
-    return;
-  }
-  if (e.target.classList.contains('js-last')) {
-    renderContentByPagination(apiService.totalPages);
-    return;
-  }
-}
-
-function renderContentByPagination(page) {
-  apiService.page = page;
-  apiService.fetchImagesByPage().then(({ data }) => renderFilms(data.results));
-  renderPagination(apiService.page, apiService.totalPages);
-}
+// ----------
 
 toggleBackdrop();
+
 function onGetInfoClick(e) {
   const filmClick = e.target.parentNode.parentNode;
   if (filmClick.nodeName !== 'LI') {
@@ -74,18 +43,6 @@ function onGetInfoClick(e) {
   const film = apiService.films.find(({ id }) => id == filmClick.id);
 
   fillModal(film);
-}
-
-function fillModal(film) {
-  refs.filmName.textContent = film.title;
-  refs.filmImage.alt = film.title;
-  refs.filmImage.src = film.poster_path;
-  refs.filmPopulation.textContent = film.popularity.toFixed(2);
-  refs.filmTittle.textContent = film.original_title;
-  refs.filmVoteFirst.textContent = film.vote_average;
-  refs.filmVoteSecond.textContent = film.vote_count;
-  refs.filmAbout.textContent = film.overview;
-  refs.filmGenres.textContent = film.genre_ids;
 }
 
 function onFormSubmit(e) {
