+++ src/js/components/modal/fillModal.js
@@ -1,0 +1,13 @@
+import { refs } from '../../references/reference.js';
+
+export function fillModal(film) {
+  refs.filmName.textContent = film.title;
+  refs.filmImage.alt = film.title;
+  refs.filmImage.src = film.poster_path;
+  refs.filmPopulation.textContent = film.popularity.toFixed(2);
+  refs.filmTittle.textContent = film.original_title;
+  refs.filmVoteFirst.textContent = film.vote_average;
+  refs.filmVoteSecond.textContent = film.vote_count;
+  refs.filmAbout.textContent = film.overview;
+  refs.filmGenres.textContent = film.genre_ids;
+}
