+++ src/js/components/smoothScroll.js
