--- src/js/spinner.js
@@ -1,11 +1,0 @@
-import { refs } from './reference.js';
-
-
-
-export function spinnerOn() {
-    refs.spinner.classList.remove('visually-hidden')
-};
-
-export function spinnerOff() {
-    refs.spinner.classList.add('visually-hidden');
-}
